Refactor for embedded search, split up files
